Added dev and prod configs to webpack
