+++ Frontend/library/package.json
@@ -2,12 +2,13 @@
     "name": "@epicgames-ps/lib-pixelstreamingfrontend-dev",
     "version": "0.0.1",
     "description": "Frontend library for Pixel Streaming",
-    "main": "dist/lib-pixelstreamingfrontend.min.js",
+    "main": "dist/lib-pixelstreamingfrontend.js",
     "types": "types/pixelstreamingfrontend.d.ts",
     "scripts": {
         "compile": "tsc --build --clean && tsc",
         "build": "npm run build-webpack",
-        "build-webpack": "npx webpack",
+        "build-webpack": "npx webpack --config webpack.prod.js",
+        "build-webpack-dev": "npx webpack --config webpack.dev.js",
         "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
         "spellcheck": "cspell \"{README.md,.github/*.md,src/**/*.ts}\""
     },
