+++ Frontend/library/webpack.common.js
@@ -5,11 +5,9 @@ const path = require('path');
 const webpack = require('webpack');
 
 module.exports = {
-    mode: 'development',
     entry: {
         index: './src/pixelstreamingfrontend.ts'
     },
-    devtool: 'inline-source-map',
     module: {
         rules: [
             {
@@ -28,15 +26,10 @@ module.exports = {
         })
     ],
     output: {
-        filename: 'lib-pixelstreamingfrontend.min.js',
         library: 'lib-pixelstreamingfrontend', // exposed variable that will provide access to the library classes
         libraryTarget: 'umd',
         path: path.resolve(__dirname, 'dist'),
-        clean: false,
+        clean: true,
         globalObject: 'this'
-    },
-    optimization: {
-        minimize: false
-    },
-    stats: 'errors-only'
+    }
 };
