+++ Frontend/library/webpack.dev.js
@@ -1,0 +1,15 @@
+// Copyright Epic Games, Inc. All Rights Reserved.
+
+const { merge } = require('webpack-merge');
+const common = require('./webpack.common.js');
+
+module.exports = merge(common, {
+    mode: 'development',
+    devtool: 'inline-source-map',
+    devServer: {
+      static: './dist',
+    },
+    output: {
+        filename: 'lib-pixelstreamingfrontend.js',
+    }
+});
